student-management/requirements.txt
Flask
pymongo
Flask-JWT-Extended
python-dotenv
